pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "agent-markitdown"
version = "0.1.0"
description = "Safe local document-to-markdown preprocessing for OpenClaw, Claude Code, Codex, Hermes, and other agents."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Threshold" }
]
keywords = ["markdown", "pdf", "docx", "agents", "openclaw", "claude-code", "codex"]
dependencies = [
  "markitdown[docx,pdf,pptx,xlsx]>=0.1.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "python-docx>=1.1.2",
  "reportlab>=4.4.1",
]

[project.scripts]
agent-markitdown = "agent_markitdown.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_markitdown"]
